requirements.txt

aiogram>=3.15.0

SQLAlchemy>=2.0.35
aiosqlite>=0.20.0
asyncpg>=0.30.0

python-dotenv>=1.0.1

apscheduler>=3.10.4
pytz>=2025.1
